+++ test/lib/production.spec.js
@@ -19,7 +19,7 @@ describe("lib/production", () => {
 
   describe("findProdInstalls", () => {
     it("throws on empty root directory with default path", () =>
-      expect(findProdInstalls()).to.be.rejectedWith("Unable to find root package.json")
+      expect(findProdInstalls()).to.be.rejectedWith("Unable to find package.json")
     );
 
     it("throws on bad package.json", async () => {
@@ -291,6 +291,94 @@ describe("lib/production", () => {
         "node_modules/foo",
         "node_modules/nested-foo"
       ]));
+    });
+
+    it("handles symlinks", async () => {
+      mock({
+        "package.json": JSON.stringify({
+          dependencies: {
+            "@scope/bar": "^1.2.3",
+            foo: "^2.3.4"
+          }
+        }),
+        node_modules: {
+          "@scope": {
+            bar: mock.symlink({
+              path: "../../lib/bar"
+            })
+          },
+          foo: {
+            "package.json": JSON.stringify({})
+          }
+        },
+        lib: {
+          bar: {
+            "package.json": JSON.stringify({
+              dependencies: {
+                baz: "^1.2.3"
+              }
+            }),
+            node_modules: {
+              baz: {
+                "package.json": JSON.stringify({})
+              }
+            }
+          }
+        }
+      });
+
+      expect(await findProdInstalls()).to.eql(normalize([
+        "node_modules/@scope/bar",
+        "node_modules/@scope/bar/node_modules/baz",
+        "node_modules/foo"
+      ]));
+    });
+
+    it("handles symlinks and monorepos", async () => {
+      mock({
+        "package.json": JSON.stringify({}),
+        packages: {
+          "my-pkg": {
+            "package.json": JSON.stringify({
+              dependencies: {
+                "@scope/bar": "^1.2.3",
+                foo: "^2.3.4"
+              }
+            })
+          }
+        },
+        node_modules: {
+          "@scope": {
+            bar: mock.symlink({
+              path: "../../lib/bar"
+            })
+          },
+          foo: {
+            "package.json": JSON.stringify({})
+          }
+        },
+        lib: {
+          bar: {
+            "package.json": JSON.stringify({
+              dependencies: {
+                baz: "^1.2.3"
+              }
+            }),
+            node_modules: {
+              baz: {
+                "package.json": JSON.stringify({})
+              }
+            }
+          }
+        }
+      });
+
+      expect(await findProdInstalls({ curPath: path.resolve("packages/my-pkg") }))
+        .to.eql(normalize([
+          "node_modules/@scope/bar",
+          "node_modules/@scope/bar/node_modules/baz",
+          "node_modules/foo"
+        ]));
     });
   });
 });